Bloco de inicializacao started
