+++ src/br/com/kaiorr/javaoo/javacore/domain/blocosinicializacao/test/AnimeTest01.java
@@ -1,0 +1,13 @@
+package br.com.kaiorr.javaoo.javacore.domain.blocosinicializacao.test;
+
+import br.com.kaiorr.javaoo.javacore.domain.blocosinicializacao.domain.Anime;
+
+public class AnimeTest01 {
+    public static void main(String[] args) {
+        Anime anime = new Anime();
+
+        for (int episodio: anime.getEpisodios()) {
+            System.out.print(episodio + " ");
+        }
+    }
+}
