+++ src/br/com/kaiorr/javaoo/javacore/domain/metodos/domain/Estudante.java
@@ -1,0 +1,14 @@
+package br.com.kaiorr.javaoo.javacore.domain.metodos.domain;
+
+public class Estudante {
+    public String nome;
+    public int idade;
+    public char sexo;
+
+    public void imprime() {
+        System.out.println("---------");
+        System.out.println(this.nome);
+        System.out.println(this.idade);
+        System.out.println(this.sexo);
+    }
+}
