+++ src/br/com/kaiorr/javaoo/javacore/domain/metodos/test/CalculadoraTest01.java
@@ -1,13 +1,11 @@
-package br.com.kaiorr.javaoo.javacore.domain.test;
+package br.com.kaiorr.javaoo.javacore.domain.metodos.test;
 
 import br.com.kaiorr.javaoo.javacore.domain.metodos.domain.Calculadora01;
 
-public class ProfessorTest01 {
+public class CalculadoraTest01 {
     public static void main(String[] args) {
         Calculadora01 calculadora = new Calculadora01();
 
-        calculadora.Soma();
-        System.out.println("-------");
-        calculadora.Subtrai();
+        calculadora.MultiplicaDoisNumeros(15, 20);
     }
 }
