+++ src/br/com/kaiorr/javaoo/javacore/domain/sobrecarga/domain/Anime.java
@@ -1,0 +1,60 @@
+package br.com.kaiorr.javaoo.javacore.domain.sobrecarga.domain;
+
+public class Anime {
+    private String nome;
+    private String tipo;
+    private int episodios;
+    private String genero;
+
+    public Anime(String nome, String tipo, int episodios, String genero) {
+        this();
+        this.nome = nome;
+        this.tipo = tipo;
+        this.episodios = episodios;
+        this.genero = genero;
+    }
+
+    public Anime() {
+        System.out.println("Detro do construtor sem argumento");
+    }
+
+    public void imprime() {
+        System.out.println(this.getNome());
+        System.out.println(this.getTipo());
+        System.out.println(this.getEpisodios());
+        System.out.println(this.getGenero());
+    }
+
+
+    public String getGenero() {
+        return genero;
+    }
+
+    public void setGenero(String genero) {
+        this.genero = genero;
+    }
+
+    public void setNome(String nome) {
+        this.nome = nome;
+    }
+
+    public String getNome() {
+        return this.nome;
+    }
+
+    public void setTipo(String tipo) {
+        this.tipo = tipo;
+    }
+
+    public void setEpisodios(int episodios) {
+        this.episodios = episodios;
+    }
+
+    public String getTipo() {
+        return this.tipo;
+    }
+
+    public int getEpisodios() {
+        return this.episodios;
+    }
+}
