--- src/br/com/kaiorr/javaoo/metodos/domain/Estudante.java
@@ -1,14 +1,0 @@
-package br.com.kaiorr.javaoo.metodos.domain;
-
-public class Estudante {
-    public String nome;
-    public int idade;
-    public char sexo;
-
-    public void imprime() {
-        System.out.println("---------");
-        System.out.println(this.nome);
-        System.out.println(this.idade);
-        System.out.println(this.sexo);
-    }
-}
