--- src/br/com/kaiorr/javaoo/metodos/test/CalculadoraTest01.java
@@ -1,11 +1,0 @@
-package br.com.kaiorr.javaoo.metodos.test;
-
-import br.com.kaiorr.javaoo.metodos.domain.Calculadora01;
-
-public class CalculadoraTest01 {
-    public static void main(String[] args) {
-        Calculadora01 calculadora = new Calculadora01();
-
-        calculadora.MultiplicaDoisNumeros(15, 20);
-    }
-}
